pyproject.toml
[project]
name = "classie-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[all]>=0.115.12",
    "httpx>=0.28.1",
    "motor>=3.7.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.11.3",
    "pymongo>=4.12.0",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.4.0",
    "uvicorn[standard]>=0.34.0",
]
